[front] refactor: rename utils/rateLater -> features/alreadyCompared
