+++ frontend/src/hooks/useLoginState.ts
@@ -8,7 +8,7 @@ import {
 import { clearSettings } from 'src/features/settings/userSettingsSlice';
 import { LoginState } from 'src/features/login/LoginState.model';
 import { isLoggedIn as isStateLoggedIn } from 'src/features/login/loginUtils';
-import { clearAllSuggested } from 'src/utils/rateLater';
+import { clearAllSuggested } from 'src/features/rateLater/alreadySuggested';
 
 export const useLoginState = () => {
   const loginState: LoginState = useAppSelector(selectLogin);
